requirement.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
Flask==1.0.2
gevent==1.3.6
greenlet==0.4.15
idna==2.7
itsdangerous==0.24
Jinja2==2.10
locustio==0.9.0
MarkupSafe==1.0
msgpack==0.5.6
pyzmq==17.1.2
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
